Added filestream GetSize function and minor bug fixes
After these changes, the first 4GiB is only mapped to HHDM address space
and not identity mapped. This is a major step required for loading
executables.

This GetSize function allows anything to be able to find out a file size
at runtime.

The reordering of user GDT segments allows system call extensions to
work later on in the project.