Added user support to scheduler, basic ELF loading and general refactoring.
The Scheduler now has the standard kernel priority and 3 different user priorities: LOW, NORMAL, HIGH. Each privilege level gets 80% more CPU time than the level below it. If there are no tasks at a certain level, the level below gets to run a task on that rotation. This allows higher priority tasks to not be slowed down by lower priority tasks that much.

System calls work as follows. Currently all 64 system calls just redirect to the same function which just prints the number and the arguments. The system call number is placed in `rax` and the arguments are parsed in `rdi`, `rsi` and `rdx`, return value is in `rax`. This is just like the System V x86_64 calling convention.

Various refactoring also occurred. Some of which to be more ISO C
compliant.