Proper file descriptor management and open, close, read, write and seek system calls
With this new file descriptor management, it is finally a lot easier to
open and read from files with the kernel. This also made it possible to
implement open, close, read, write and seek system calls.

Some system call entry changes occurred to allow for interrupts to occur
during system calls without messing up the kernel GS base.

NOTE: git says that the old HAL/vfs.hpp file was renamed to a LibC
source file. This is not the case.