Improved printf and panic screen
This new printf functionality allows proper width specification, 2 flags
and keeping track of the amount of characters printed. This is making
printf get a lot closer to following the spec.

printf's new functionality was used to help improve the panic screen and
make it easier to understand. Other changes to the panic screen also
occurred.