Eternal kernel heap and address space reservation
This eternal kernel heap allows the old kernel heap to be slightly
smaller because big items like the kernel symbol table are now being
eternally allocated.

The new virtual address space reservation support allows memory to be
allocated without having a physical memory backing. Currently,
allocation of the memory must be requested for physical memory to be
allocated for it. In future, page fault handling will mean that memory
will be automatically given physical memory when it is used for the
first time. This new future is also a large step in implementing
swapping.