Page allocation no longer must be physically continuous and other changes
This removed requirement allows for better allocation of large sizes,
but could cause physical memory fragmentation. With this change, proper
page remapping functions were implemented on an architecture specific
level, instead of just unmapping and then mapping again with the new
permissions. Out of memory detection was also added. This is another
step towards swapping.

Other changes include using the host systems OVMF install instead of the
old one that has been in the repository for a long time. The variables
are copied over and can be updated by running a full clean before
building. The code is used as read-only and loaded directory from the
host systems /usr/share/edk2 folder.