Remove the usage of WorldOS namespace everywhere in the kernel
Originally, most architecture in-dependant kernel related code was
inside the WorldOS namespace. Eventually this was not being done
anymore. With this commit, all usages of the WorldOS namepace are now
gone. It is indented that it will never be used in the future.

Starting with this commit, a changelog will only be written when many
changes occurred. As this was only one big change, there is no point.