PS/2 Keyboard Input support
This allows for keyboard input on most virtual machines and some real
hardware. This has been a feature that has been needed for a while, and
has been finally implemented. This will make it easier for creating a
shell.

Many graphical bugs were discovered in the process of implementing this.
Proper console scrolling now works and wrapping around to the next line
also now works. Backspace behaves correctly and removes all pixels that
it is supposed to.