exec system call
This means that userland programs can now execute other programs. This
is a massive milestone in the project. This was implemented in a way
where executing programs in kernel-land is now much easier.

Some improvements to time functions in the kernel were also made to
improve code readablity and correctness of leap year checking.