Kernel Lib, better toolchain building, and LibC header improvements
The new kernel library, contains all the kernel variants of the LibC
code. This used to be part of the main kernel code, but it was moved to
make the kernel code cleaner. This has been long overdue, as it was
making the kernel code increasingly messy.

The toolchain build script now uses the WorldOS variants of the GCC and
binutils-gdb repositories for building the toolchain. This allows for
much faster download, as the download, decompressing and patches is all
done as one step.

More ISO C11 compliant header files have been added to LibC. An extra
errno code, which is identical to another, was also added.