Improved seek system call and minor toolchain changes
The seek system call now takes a third parameter which detemines the
starting point of the seek. This can either be the start of the stream,
the current position in the stream or the end of the stream. This now
means that fseek should now be ISO C compliant.

GCC was updated to 13.2.1 and the toolchain prefix was moved to
`toolchain/local` to avoid potential conflicts with other toolchains.