ls, directory stream support, kinput bug fixes
Opening directory support was added to the open system call and a
getdirents system call was implemented to allow for suoport for ls. Many
VFS functions had to either be changed or added to allow for this to
happen. The seek system call doesn't support directory streams yet.

The stat system call now exposes the type of an inode. This can either
be a file, directory or symbolic link.

Many changes had to be made to the Buffer and KeyboardInput class to
allow for keyboard input to work correctly due to interrupt-driven
input. Both class had to have changes made so they could handle writes
occurring either in the middle of reads or writes.