Updated kernel heap, added a heap to LibC, README cleanup, run-utils fixes.
We now have a fully-custom kernel heap which performs fairly well. In
the process of implementing this, a userspace heap was designed and
implemented in LibC.

LibC now supports the _Exit function which is very similar to exit,
except it does not do any cleanup.

The stat program now prints the file type. Ideally this should have been
in yesterdays big commit, but it wasn't.

The toolchain path was updated in all the run-utils. The toolchain build
requirements were updated in the README to include git.

The run command in README.md was updated. This should have been updated
when the ovmf changes occurred, but I forgot to.