Working directory support
All threads have their own working directory, and all processes have
a default working directory which is inherited from the parent process,
and can be changed with by the main thread. This is currently unsupported
on file, folder and symlink creation. This came with chdir and fchdir
system calls. It also meant that many VFS-related functions needed to be
changed.

The buildsymboltable utility was also updated due to unknown issues. It
now just takes the output from nm into stdin instead of trying to
execute it and create pipes.