Signalling support and better exception handling.
Basic signalling support has been implemented. This means a process
can send a signal to another process with a priority less than or equal
to it's priority. These signals can be caught by the process and handled.
Signals are always handled on the main thread of a process. This came
with three the system calls: onsignal, sendsig and sigreturn. When a
signal handler registered by onsignal is called, some code injection
occurs so that the sigreturn syscall is called to perform clean-up after
the signal handler and restore the thread state.

Thread specfic exception handling was also implemented. When an exception
occurs in a process, the main thread is sent an appropriate signal. On
x86_64, divide by zero and floating point exceptions send a SIGFPE
signal, Invalid Opcode sends a SIGILL signal, and page faults and
general protection faults send a SIGSEGV signal.

Some LibC functions were either added or updated to allow for handling
of this new signalling support.

Better page fault handling was also implemented.