Update to limine 6.x, LibC improvements, memset fixed
We now use limine 6.x as a bootloader. This also means we now use limine
protocol revision 1.

All of the ISO C ctype.h function have now been implemented in LibC.
There was also a note added about how gets is not supported in C11 or
newer.

The memset function in both the kernel and LibC has now been fixed.