Paging cleanup and minor build improvements.
The page mapping/unmapping functions now do not map freshly allocated
tables as there are in the HHDM. Also removed the usage of the old
deprecated fast_memset function in this code. Most of page tables that
are included in the kernel are now gone as they are unneeded. The only
exception to this is the kernel PML4 array.

The RTC GetWeekDay function now uses the correct formula in order to get
the correct day for all dates. When the RTC time is printed on init, the
hour, minute, and second fields now have a minimum width of 2 digits.

Some minor build system improvements were also made so clangd can be
used.